requirements.txt
flask==3.0.0
google-generativeai==0.8.3
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0
openpyxl==3.1.2